Finish adding basic type hints to everything.
Finish adding basic type hints to everything.

Several of the modules need refactoring and specific type aliases for
interfaces.  This is currently in progress in several other branches.

A big purpose of this branch is to get mypy working with CI, so future
refactors are less likely to break the code.

Break break break.